Moved Jobs methods to a provider.
