+++ src/main/java/nl/aurorion/blockregen/provider/WorldGuardProvider.java
@@ -1,4 +1,4 @@
-package nl.aurorion.blockregen;
+package nl.aurorion.blockregen.provider;
 
 import com.sk89q.worldedit.bukkit.BukkitAdapter;
 import com.sk89q.worldguard.LocalPlayer;
@@ -8,6 +8,7 @@ import com.sk89q.worldguard.protection.flags.Flags;
 import com.sk89q.worldguard.protection.flags.StateFlag;
 import com.sk89q.worldguard.protection.regions.RegionContainer;
 import com.sk89q.worldguard.protection.regions.RegionQuery;
+import nl.aurorion.blockregen.BlockRegen;
 import org.bukkit.Location;
 import org.bukkit.entity.Player;
